// src/generated/coc-api/api.ts
export interface BadgeUrls {
    'small': string;
    'medium': string;
    'large': string;
}
// ...
export interface ClanWar {
    'clan'?: WarClan;
    'teamSize'?: number;
    'attacksPerMember'?: number;
    'opponent'?: WarClan;
    'startTime'?: string;
    'state'?: WarState | null;
    'endTime'?: string;
    'preparationStartTime'?: string;
}
// ...
export interface ClanWarMember {
    'tag'?: string;
    'name'?: string;
    'mapPosition'?: number;
    'townhallLevel'?: number;
    'opponentAttacks'?: number;
    'bestOpponentAttack'?: ClanWarAttack | null;
    'attacks'?: Array<ClanWarAttack> | null;
}
// ...
export interface WarClan {
    'destructionPercentage'?: number;
    'tag'?: string;
    'name'?: string;
    'badgeUrls'?: BadgeUrls;
    'clanLevel'?: number;
    'attacks'?: number;
    'stars'?: number;
    'expEarned'?: number;
    'members'?: Array<ClanWarMember>;
}
// ...
export const WarState = {
    NotInWar: 'notInWar',
    Preparation: 'preparation',
    InWar: 'inWar',
    WarEnded: 'warEnded'
} as const;

export type WarState = typeof WarState[keyof typeof WarState];